refactor trust basis test layout
